Separate devicetree overlay directory for each kernel (Closes: #1095811)
A devicetree overlay may differ between kernel versions and is not
necessarily compatible between different kernel versions, same as the
devicetree itself. Therefore use dedicated directories for devicetree
overlays, in the same way as for the devicetree files.

The update script first checks a version specific directory. If the
directory does not exist, a second check is done in the directory
without version.

Signed-off-by: Markus Burri <markus.burri@mt.com>